# mapper cache: Use string concatenation rather than Sprintf for formatKey