While doing some profiling for something else, I noticed `fmt.Sprintf` in the hot path for cache operations. Changing this to string concatenation gives some improvement in performance.

benchmarked using `go test -bench=BenchmarkGlob100RulesMultipleCapturesCached -benchtime=10s -benchmem`

output from benchcmp:

```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkGlob100RulesMultipleCapturesCached-12     280           158           -43.57%

benchmark                                          old allocs     new allocs     delta
BenchmarkGlob100RulesMultipleCapturesCached-12     4              2              -50.00%

benchmark                                          old bytes     new bytes     delta
BenchmarkGlob100RulesMultipleCapturesCached-12     96            64            -33.33%
```